% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iotdataplane_operations.R
\name{iotdataplane_publish}
\alias{iotdataplane_publish}
\title{Publishes state information}
\usage{
iotdataplane_publish(topic, qos, payload)
}
\arguments{
\item{topic}{[required] The name of the MQTT topic.}

\item{qos}{The Quality of Service (QoS) level.}

\item{payload}{The state information, in JSON format.}
}
\description{
Publishes state information.
}
\details{
For more information, see \href{http://docs.aws.amazon.com/iot/latest/developerguide/protocols.html#http}{HTTP Protocol}
in the \emph{AWS IoT Developer Guide}.
}
\section{Request syntax}{
\preformatted{iotdataplane$publish(
  topic = "string",
  qos = 123,
  payload = raw
)
}
}

\keyword{internal}
